backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.0
python-jose==3.3.0
httpx==0.25.2
